ict assignment Q9
Write a C program with a function named isEven to take an integer value and return 1 if the number is even and 0 if it is not. In the main function, use the isEven function to check whether a given number is even or not. If the number is even, display “It is an even number.”, otherwise display “It is an odd number.”